+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.4.0
+* Fix encoding and decoding of dynamic types (https://github.com/poanetwork/ex_abi/pull/34)
+* Allow to decoded function outputs (https://github.com/poanetwork/ex_abi/pull/36)
+* Parse array of tuples in the specification (https://github.com/poanetwork/ex_abi/pull/37)
+# 0.3.2
+* Fix array/tuple decoding (https://github.com/poanetwork/ex_abi/pull/32)
+# 0.3.1
+* ABI parsing tuple type (https://github.com/poanetwork/ex_abi/pull/29)
+* ABI encoding array type fix (https://github.com/poanetwork/ex_abi/pull/28)
+* Elixir version bump: 1.10.2 (https://github.com/poanetwork/ex_abi/pull/27)
 # 0.3.0
 * Fix encoding/decoding of dynamic size types (https://github.com/poanetwork/ex_abi/pull/24)
 # 0.2.2